Move gas update test to test_gas_update.rs
