+++ crates/starknet-devnet/tests/common/mod.rs
@@ -2,5 +2,6 @@ pub mod background_anvil;
 pub mod background_devnet;
 pub mod constants;
 pub mod errors;
+pub mod fees;
 pub mod reqwest_client;
 pub mod utils;
