+++ crates/starknet-devnet/tests/test_simulate_transactions.rs
@@ -4,9 +4,7 @@ mod simulation_tests {
     use std::sync::Arc;
 
     use serde_json::json;
-    use starknet_core::constants::{
-        CAIRO_0_ACCOUNT_CONTRACT_HASH, DEVNET_DEFAULT_GAS_PRICE, ETH_ERC20_CONTRACT_ADDRESS,
-    };
+    use starknet_core::constants::{CAIRO_0_ACCOUNT_CONTRACT_HASH, ETH_ERC20_CONTRACT_ADDRESS};
     use starknet_core::utils::exported_test_utils::dummy_cairo_0_contract_class;
     use starknet_rs_accounts::{
         Account, AccountFactory, Call, ExecutionEncoding, OpenZeppelinAccountFactory,
@@ -24,68 +22,12 @@ mod simulation_tests {
     use starknet_rs_signers::Signer;
 
     use crate::common::background_devnet::BackgroundDevnet;
-    use crate::common::constants::{CAIRO_1_CONTRACT_PATH, CHAIN_ID, INTEGRATION_SEPOLIA_HTTP_URL};
+    use crate::common::constants::{CAIRO_1_CONTRACT_PATH, CHAIN_ID};
+    use crate::common::fees::{assert_difference_if_validation, assert_fee_in_resp_at_least_equal};
     use crate::common::utils::{
         get_deployable_account_signer, get_flattened_sierra_contract_and_casm_hash,
         iter_to_hex_felt, to_hex_felt, to_num_as_hex,
     };
-
-    fn extract_overall_fee(simulation_result: &serde_json::Value) -> u128 {
-        let fee_hex = simulation_result["fee_estimation"]["overall_fee"].as_str().unwrap();
-        let fee_hex_stripped = fee_hex.strip_prefix("0x").unwrap();
-        u128::from_str_radix(fee_hex_stripped, 16).unwrap()
-    }
-
-    /// Calculating the fee of a transaction depends on fee weights of vm resources and the resource
-    /// usage. The fee from vm usage is the heaviest product of cairo resource usage multiplied
-    /// by cairo resource fee cost, rounded down to the nearest integer. There is a
-    /// possibility that although some of the resources present in the resp_no_flags are not in
-    /// resp_skip_validation, this doesnt mean that the fee will be higher, for example:
-    /// the fee from vm usage is determined from the `steps` and the transaction that skips
-    /// validation might have less steps, but due to the rounding if the multiplied product
-    /// falls in the range (7,8] both will be rounded to 7
-    fn assert_fee_in_resp_at_least_equal(
-        resp_no_flags: &serde_json::Value,
-        resp_skip_validation: &serde_json::Value,
-    ) {
-        let no_flags_fee = extract_overall_fee(resp_no_flags);
-        let skip_validation_fee = extract_overall_fee(resp_skip_validation);
-        assert!(no_flags_fee.ge(&skip_validation_fee));
-    }
-
-    fn assert_difference_if_validation(
-        resp_no_flags: &serde_json::Value,
-        resp_skip_validation: &serde_json::Value,
-        expected_contract_address: &str,
-        should_skip_fee_invocation: bool,
-    ) {
-        let no_flags_trace = &resp_no_flags["transaction_trace"];
-        assert_eq!(
-            no_flags_trace["validate_invocation"]["contract_address"].as_str().unwrap(),
-            expected_contract_address
-        );
-        assert!(no_flags_trace["state_diff"].as_object().is_some());
-
-        let skip_validation_trace = &resp_skip_validation["transaction_trace"];
-        assert!(skip_validation_trace["validate_invocation"].as_object().is_none());
-        assert!(skip_validation_trace["state_diff"].as_object().is_some());
-
-        assert_eq!(
-            skip_validation_trace["fee_transfer_invocation"].as_object().is_none(),
-            should_skip_fee_invocation
-        );
-        assert_eq!(
-            no_flags_trace["fee_transfer_invocation"].as_object().is_none(),
-            should_skip_fee_invocation
-        );
-
-        assert!(
-            no_flags_trace["execution_resources"]["steps"].as_u64().unwrap()
-                > skip_validation_trace["execution_resources"]["steps"].as_u64().unwrap()
-        );
-
-        assert_fee_in_resp_at_least_equal(resp_no_flags, resp_skip_validation);
-    }
 
     #[tokio::test]
     async fn simulate_declare_v1() {
@@ -231,238 +173,6 @@ mod simulation_tests {
             resp_skip_validation,
             &sender_address_hex,
             max_fee == FieldElement::ZERO,
-        );
-    }
-
-    async fn update_gas_scenario(devnet: BackgroundDevnet, expected_chain_id: &str) {
-        // get account
-        let (signer, account_address) = devnet.get_first_predeployed_account().await;
-        let account = SingleOwnerAccount::new(
-            devnet.clone_provider(),
-            signer,
-            account_address,
-            FieldElement::from_hex_be(expected_chain_id).unwrap(),
-            ExecutionEncoding::New,
-        );
-
-        // get class
-        let (flattened_contract_artifact, casm_hash) =
-            get_flattened_sierra_contract_and_casm_hash(CAIRO_1_CONTRACT_PATH);
-
-        let max_fee = FieldElement::ZERO;
-        let nonce = FieldElement::from_hex_be("0x0").unwrap();
-
-        let signature = account
-            .declare(Arc::new(flattened_contract_artifact.clone()), casm_hash)
-            .max_fee(max_fee)
-            .nonce(nonce)
-            .prepared()
-            .unwrap()
-            .get_declare_request(false)
-            .await
-            .unwrap()
-            .signature;
-
-        let signature_hex: Vec<String> = iter_to_hex_felt(&signature);
-        let sender_address_hex = to_hex_felt(&account_address);
-        let get_params = |simulation_flags: &[&str]| -> serde_json::Value {
-            json!({
-                "block_id": "latest",
-                "simulation_flags": simulation_flags,
-                "transactions": [
-                    {
-                        "type": "DECLARE",
-                        "sender_address": sender_address_hex,
-                        "compiled_class_hash": to_hex_felt(&casm_hash),
-                        "max_fee": to_hex_felt(&max_fee),
-                        "version": "0x2",
-                        "signature": signature_hex,
-                        "nonce": to_num_as_hex(&nonce),
-                        "contract_class": flattened_contract_artifact,
-                    }
-                ]
-            })
-        };
-
-        let chain_id = &devnet.send_custom_rpc("starknet_chainId", json!({})).await.unwrap();
-        assert_eq!(chain_id, expected_chain_id);
-
-        let params_no_flags = get_params(&[]);
-        let resp_no_flags = &devnet
-            .send_custom_rpc("starknet_simulateTransactions", params_no_flags.clone())
-            .await
-            .unwrap()[0];
-        assert_eq!(resp_no_flags["fee_estimation"]["gas_price"], "0x174876e800");
-        assert_eq!(resp_no_flags["fee_estimation"]["data_gas_price"], "0x174876e800");
-        assert_eq!(resp_no_flags["fee_estimation"]["overall_fee"], "0x7352ecf52000");
-
-        let params_skip_validation = get_params(&["SKIP_VALIDATE"]);
-        let resp_skip_validation = &devnet
-            .send_custom_rpc("starknet_simulateTransactions", params_skip_validation.clone())
-            .await
-            .unwrap()[0];
-        assert_eq!(resp_skip_validation["fee_estimation"]["gas_price"], "0x174876e800");
-        assert_eq!(resp_skip_validation["fee_estimation"]["data_gas_price"], "0x174876e800");
-        assert_eq!(resp_skip_validation["fee_estimation"]["overall_fee"], "0x733ba47e3800");
-
-        assert_difference_if_validation(
-            resp_no_flags,
-            resp_skip_validation,
-            &sender_address_hex,
-            max_fee == FieldElement::ZERO,
-        );
-
-        let gas_update = json!({
-            "gas_price_wei": 9e18 as u128,
-            "data_gas_price_wei": 8e18 as u128,
-            "gas_price_strk": 7e18 as u128,
-            "data_gas_price_strk": 6e18 as u128,
-            "generate_block": true,
-        });
-        let updated_gas =
-            &devnet.send_custom_rpc("devnet_updateGas", gas_update.clone()).await.unwrap();
-        assert_eq!(updated_gas, &gas_update);
-
-        let chain_id = &devnet.send_custom_rpc("starknet_chainId", json!({})).await.unwrap();
-        assert_eq!(chain_id, expected_chain_id);
-
-        let resp_no_flags = &devnet
-            .send_custom_rpc("starknet_simulateTransactions", params_no_flags)
-            .await
-            .unwrap()[0];
-        assert_eq!(resp_no_flags["fee_estimation"]["gas_price"], "0x7ce66c50e2840000");
-        assert_eq!(resp_no_flags["fee_estimation"]["data_gas_price"], "0x6f05b59d3b200000");
-        assert_eq!(resp_no_flags["fee_estimation"]["overall_fee"], "0x2603cc779e46ad00000");
-
-        let resp_skip_validation = &devnet
-            .send_custom_rpc("starknet_simulateTransactions", params_skip_validation)
-            .await
-            .unwrap()[0];
-        assert_eq!(resp_skip_validation["fee_estimation"]["gas_price"], "0x7ce66c50e2840000");
-        assert_eq!(resp_skip_validation["fee_estimation"]["data_gas_price"], "0x6f05b59d3b200000");
-        assert_eq!(resp_skip_validation["fee_estimation"]["overall_fee"], "0x25fbfe10d93884c0000");
-
-        assert_difference_if_validation(
-            resp_no_flags,
-            resp_skip_validation,
-            &sender_address_hex,
-            max_fee == FieldElement::ZERO,
-        );
-    }
-
-    #[tokio::test]
-    async fn update_gas() {
-        let devnet = BackgroundDevnet::spawn().await.expect("Could not start Devnet");
-
-        update_gas_scenario(devnet, "0x534e5f5345504f4c4941").await;
-    }
-
-    #[tokio::test]
-    async fn update_gas_fork() {
-        let cli_args: [&str; 2] = ["--fork-network", INTEGRATION_SEPOLIA_HTTP_URL];
-        let fork_devnet = BackgroundDevnet::spawn_with_additional_args(&cli_args).await.unwrap();
-
-        update_gas_scenario(fork_devnet, "0x534e5f494e544547524154494f4e5f5345504f4c4941").await;
-    }
-
-    #[tokio::test]
-    async fn update_gas_check_blocks() {
-        let devnet = BackgroundDevnet::spawn().await.expect("Could not start Devnet");
-        devnet.create_block().await.unwrap();
-        let latest_block = devnet.get_latest_block_with_txs().await.unwrap();
-        assert_eq!(latest_block.block_number, 1);
-        assert_eq!(
-            latest_block.l1_gas_price.price_in_wei,
-            FieldElement::from(u128::from(DEVNET_DEFAULT_GAS_PRICE))
-        );
-        assert_eq!(
-            latest_block.l1_gas_price.price_in_fri,
-            FieldElement::from(u128::from(DEVNET_DEFAULT_GAS_PRICE))
-        );
-
-        let wei_price_first_update = 9e18 as u128;
-        let strk_price_first_update = 7e18 as u128;
-        let gas_update = json!({
-            "gas_price_wei": wei_price_first_update,
-            "data_gas_price_wei": 8e18 as u128,
-            "gas_price_strk": strk_price_first_update,
-            "data_gas_price_strk": 6e18 as u128,
-            "generate_block": false,
-        });
-        let updated_gas =
-            &devnet.send_custom_rpc("devnet_updateGas", gas_update.clone()).await.unwrap();
-        assert_eq!(updated_gas, &gas_update);
-
-        let latest_block = devnet.get_latest_block_with_txs().await.unwrap();
-        assert_eq!(latest_block.block_number, 1);
-
-        let pending_block = devnet.get_pending_block_with_tx_hashes().await.unwrap();
-        assert_eq!(
-            pending_block.l1_gas_price.price_in_wei,
-            FieldElement::from(u128::from(DEVNET_DEFAULT_GAS_PRICE))
-        );
-        assert_eq!(
-            pending_block.l1_gas_price.price_in_fri,
-            FieldElement::from(u128::from(DEVNET_DEFAULT_GAS_PRICE))
-        );
-
-        devnet.create_block().await.unwrap();
-
-        let pending_block = devnet.get_pending_block_with_tx_hashes().await.unwrap();
-        assert_eq!(
-            pending_block.l1_gas_price.price_in_wei,
-            FieldElement::from(wei_price_first_update)
-        );
-        assert_eq!(
-            pending_block.l1_gas_price.price_in_fri,
-            FieldElement::from(strk_price_first_update)
-        );
-
-        let latest_block = devnet.get_latest_block_with_txs().await.unwrap();
-        assert_eq!(latest_block.block_number, 2);
-        assert_eq!(
-            latest_block.l1_gas_price.price_in_wei,
-            FieldElement::from(u128::from(DEVNET_DEFAULT_GAS_PRICE))
-        );
-        assert_eq!(
-            latest_block.l1_gas_price.price_in_fri,
-            FieldElement::from(u128::from(DEVNET_DEFAULT_GAS_PRICE))
-        );
-
-        let wei_price_second_update = 8e18 as u128;
-        let strk_price_second_update = 6e18 as u128;
-        let gas_update_with_new_block = json!({
-            "gas_price_wei": wei_price_second_update,
-            "data_gas_price_wei": 7e18 as u128,
-            "gas_price_strk": strk_price_second_update,
-            "data_gas_price_strk": 5e18 as u128,
-            "generate_block": true,
-        });
-        let updated_gas = &devnet
-            .send_custom_rpc("devnet_updateGas", gas_update_with_new_block.clone())
-            .await
-            .unwrap();
-        assert_eq!(updated_gas, &gas_update_with_new_block);
-
-        let latest_block = devnet.get_latest_block_with_txs().await.unwrap();
-        assert_eq!(latest_block.block_number, 3);
-        assert_eq!(
-            latest_block.l1_gas_price.price_in_wei,
-            FieldElement::from(wei_price_first_update)
-        );
-        assert_eq!(
-            latest_block.l1_gas_price.price_in_fri,
-            FieldElement::from(strk_price_first_update)
-        );
-
-        let pending_block = devnet.get_pending_block_with_tx_hashes().await.unwrap();
-        assert_eq!(
-            pending_block.l1_gas_price.price_in_wei,
-            FieldElement::from(wei_price_second_update)
-        );
-        assert_eq!(
-            pending_block.l1_gas_price.price_in_fri,
-            FieldElement::from(strk_price_second_update)
         );
     }
