+++ website/docs/gas.md
@@ -3,6 +3,7 @@
 The gas fees update endpoint allows users to modify the current gas prices on a running Devnet. This feature is particularly useful for testing purposes and for adjustments needed after forking to align with the forked network's gas prices.
 
 ## JSON-RPC Request
+
 The following JSON-RPC request can be used to update gas prices:
 
 ```
@@ -21,6 +22,7 @@ JSON-RPC
 ```
 
 ## Response
+
 The expected response from the server will mirror the request parameters, confirming the updated gas prices:
 
 ```
